Removed utils.py as the file became redundant - Everything is now in mkalias.py
