+++ mkalias_cli/mkalias.py
@@ -9,11 +9,15 @@ import logging
 import os
 import sys
 
+import osascript
 from setuptools_scm import get_version
 
 from mkalias_cli import strings
-from mkalias_cli import utils
 
+#  Get version info
+version = get_version()
+
+# Setup Logging
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
@@ -22,12 +26,10 @@ ch.setLevel(logging.DEBUG)
 
 logger.addHandler(ch)
 
-version = get_version()
-
 
 def parse_args():
     """
-    Function to setup and hold argument parser
+    Function to setup argument parser
     :return: parser.parse_args() object
     """
     parser = argparse.ArgumentParser(prog='mkalias',
@@ -42,31 +44,58 @@ def parse_args():
     return parser.parse_args()
 
 
+class Alias:
+    #  Constants used for this Class
+    CMD_STRING = 0
+    CODE = 1
+    OUT = 2
+    ERROR = 3
+
+    @staticmethod
+    def create_alias(source, destination, name=None) -> tuple:
+        """
+        Creates and runs the AppleScript required to create the alias
+        :param source: File or Directory to create an alias of
+        :param destination: Destination directory of the new alias
+        :param name: Name of new alias - OPTIONAL
+        :return: tuple containing the AppleScript Created, Code, Output of AppleScript, and Errors - in that order
+        """
+
+        if name is None:
+            cmd_string = 'tell application "Finder" to make alias file to POSIX file "{}" at POSIX file "{}"' \
+                .format(source, destination)
+        else:
+            cmd_string = 'tell application "Finder" to make alias file to POSIX file "{}" at POSIX file "{}"' \
+                         ' with properties {{name:"{}"}}'.format(source, destination, name)
+
+        code, out, error = osascript.run(cmd_string)
+
+        return cmd_string, code, out, error
+
+
 def main():
     args = parse_args()
 
     source = os.path.abspath(args.source)
     destination = os.path.abspath(args.destination)
 
-    #  TODO: Check for symbolic / hard links
-    if not utils.Path.check_path(source):
+    #  TODO: Check for symbolic / hard links if necessary
+    if not os.path.exists(source):
         logger.error(strings.PATH_NOT_FOUND.format(source))
         sys.exit(1)
-    elif not utils.Path.check_path(destination):
+    elif not os.path.exists(destination):
         logger.error(strings.PATH_NOT_FOUND.format(destination))
         sys.exit(1)
 
-    #  TODO: Make this better?
     if args.alias_name:
-        create_alias_output = utils.Alias.create_alias(source, destination, args.alias_name)
+        create_alias_output = Alias.create_alias(source, destination, args.alias_name)
     else:
-        create_alias_output = utils.Alias.create_alias(source, destination)
-        # create_alias_output = utils.Alias.create_alias(source, destination)
+        create_alias_output = Alias.create_alias(source, destination)
 
-    logger.info(create_alias_output[utils.Alias.CMD_STRING])
-    logger.debug(create_alias_output[utils.Alias.CODE])
-    logger.debug(create_alias_output[utils.Alias.OUT])
-    logger.error(create_alias_output[utils.Alias.ERROR])
+    logger.info(create_alias_output[Alias.CMD_STRING])
+    logger.debug(create_alias_output[Alias.CODE])
+    logger.debug(create_alias_output[Alias.OUT])
+    logger.error(create_alias_output[Alias.ERROR])
 
     logging.shutdown()
     sys.exit(0)  # exit gracefully
